Fix: ensure creator of task are the one deleting it

+++ server/controllers/task.controller.js
@@ -1,10 +1,24 @@
+import jwt from 'jsonwebtoken'
 import { Task } from '../models/Task.js'
 import { User } from '../models/User.js'
+import { decodeToken } from '../utils/generateToken.js'
+
 
 export const getAllTasks = async (req, res) => {
     try {
-        const tasks = await Task.find({}).sort({ createdAt: -1 })
-        res.status(200).json(tasks)
+        const token = decodeToken(req)
+        const userId = jwt.verify(token, process.env.SECRET).id
+        if (!userId) {
+            return res.status(400).json({message: 'invalid token'})
+        }
+        const user = await User.findById(userId)
+        if(!user){
+            return res.status(404).json({message: 'invalid token'})
+        }
+        const userTasks = await Task.find({user: userId}).sort({ createdAt: -1 })
+        
+        // const tasks = await Task.find({}).sort({ createdAt: -1 })
+        res.status(200).json(userTasks)
     } catch (error) {
         console.log(error)
         res.status(500).json({message: "Error fetching all tasks"})
@@ -13,12 +27,18 @@ export const getAllTasks = async (req, res) => {
 
 export const createTask = async (req, res) => {
     try {
+        const token = decodeToken(req)
         const { task, dueDate } = req.body
-        console.log(task, dueDate)
+        const userId = jwt.verify(token, process.env.SECRET).id
+        if (!userId) {
+            return res.status(400).json({message: 'invalid token'})
+        }
+        
         if (!task || task.trim() === '') {
             return res.status(400).json("Task cannot be empty")
         }
         const newTask = new Task({
+            user: userId,
             text: task.trim(),
             ...(dueDate && {dueDate})
         })
@@ -32,6 +52,11 @@ export const createTask = async (req, res) => {
 
 export const editTask = async (req, res) => {
     try {
+        const token = decodeToken(req)
+        const userId = jwt.verify(token, process.env.SECRET).id
+        if (!userId) {
+            return res.status(400).json({message: 'invalid token'})
+        }
         const { id } = req.params
         const { task, dueDate, completed } = req.body
         const checkExitence = await Task.findById(id) 
@@ -41,6 +66,9 @@ export const editTask = async (req, res) => {
         
         if (!task || task.trim() === '') {
             return res.status(400).json({message: "task field cannot be empty"})
+        }
+        if (checkExitence.user.toString() !== userId) {
+            return res.status(403).json({message: 'Unauthorized'})
         }
         const editedTask = {
             text: task.trim()
@@ -61,11 +89,18 @@ export const editTask = async (req, res) => {
 
 export const toggleCompleteTask = async (req, res) => {
     try {
+        const token = decodeToken(req)
+        const userId = jwt.verify(token, process.env.SECRET).id
+        if (!userId) {
+            return res.status(400).json({message: 'invalid token'})
+        }
         const {id} = req.params
-        console.log(id);
         const checkExistence = await Task.findById(id)
         if (!checkExistence) {
             return res.status(404).json({message: `No task with id ${id}`})
+        }
+        if (checkExistence.user.toString() !== userId) {
+            return res.status(403).json({message: 'Unauthorized'})
         }
         const toggleCompletion = await Task.findByIdAndUpdate(id, {completed : !checkExistence.completed}, {new: true})
         res.status(200).json(toggleCompletion)
@@ -77,15 +112,23 @@ export const toggleCompleteTask = async (req, res) => {
 
 export const deleteTask = async (req, res) => {
     try {
+        const token = decodeToken(req)
+        const userId = jwt.verify(token, process.env.SECRET).id
+        if (!userId) {
+            return res.status(400).json({message: 'invalid token'})
+        }
         const { id } = req.params
         const checkExitence = await Task.findById(id)
         if(!checkExitence){
             return res.status(404).json({message: `task with id ${id} not found`})
         }
+        if (checkExitence.user.toString() !== userId) {
+            return res.status(403).json({message: 'Unauthorized'})
+        }
         await Task.findByIdAndDelete(id)
         res.status(200).json({message: "task deleted successfully"})
     } catch (error) {
         console.log(error)
-        res.status(500).json({message: "error deleting post"})
+        res.status(500).json({message: "error deleting task"})
     }
 }
